ADM-373.1 fix buildKite failed issue

+++ backend/src/main/java/heartbeat/client/dto/codebase/github/LeadTime.java
@@ -1,12 +1,9 @@
 package heartbeat.client.dto.codebase.github;
 
-import heartbeat.client.dto.pipeline.buildkite.DeployInfo;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.time.Instant;
 
 @Data
 @Builder
@@ -16,20 +13,20 @@ public class LeadTime {
 
 	private String commitId;
 
-	private double prCreatedTime;
+	private long prCreatedTime;
 
-	private double prMergedTime;
+	private long prMergedTime;
 
-	private double firstCommitTimeInPr;
+	private long firstCommitTimeInPr;
 
-	private double jobFinishTime;
+	private long jobFinishTime;
 
-	private double pipelineCreateTime;
+	private long pipelineCreateTime;
 
-	private double prDelayTime;
+	private long prDelayTime;
 
-	private double pipelineDelayTime;
+	private long pipelineDelayTime;
 
-	private double totalTime;
+	private long totalTime;
 
 }
